refactor: :recycle: move logging to new file
