+++ libft/src/ysystem/ysyscall.c
@@ -6,7 +6,7 @@
 /*   By: youkim < youkim@student.42seoul.kr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/20 17:40:48 by youkim            #+#    #+#             */
-/*   Updated: 2021/11/23 11:43:03 by youkim           ###   ########.fr       */
+/*   Updated: 2021/11/23 13:29:50 by youkim           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,18 +16,6 @@
 int	ywrite(int fd, char *str)
 {
 	return (write(fd, str, ystrlen(str)));
-}
-
-//	log at STDERR, but does not kill process
-void	ywarn(char *where, char *warn)
-{
-	char	*msg;
-
-	msg = new_ystrjoin((char *[]) \
-		{"[Warning @", where, "]\n    ", warn, "\n", NULL} \
-	);
-	ywritecolor(2, msg, YEL);
-	del_ystr(msg);
 }
 
 //	same as write but with colors
